backend/requirements.txt
django==5.1.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-environ==0.13.0
psycopg2-binary==2.9.11
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
weasyprint==68.1
gunicorn==25.1.0
whitenoise==6.11.0
pillow==12.1.1
